Set a mock API system for development
Why? The backend API isn't ready, and super realiable yet.

Details:

Relly on the testing lib `fetch-mock`, a mock of the fetch function.
When NODE_ENV is not production, `fetch` is globally replaced by
`fetch-mock`.

`fetch-mock` waits for match on a http verb and uri couple. When such
a match happen, the `fetch-mock` returns a specified js constant.
See src/mockedAPI/routes.js for an example with the data fetched by
the PredefinedCorpus component.