Add dynamic filters to <CustomCorpus />
Note: the interface with the filter endpoint remained to be
more precisely defined to be useful. The mock I created for
testing relies on a few assumptions. Once the requirement are
clear, this code would benefit from being typed.

Note 2: the logic behind these dynamic filters is complex, would
ideally have to be tested extensively, at least at a component
level with Jest.