Store the user login status in the browser state
Why?
The frontend needs to know if the user is logged-in to
adapt the navbar accordingly. Yet, the PHPSESSID set by
the backend doesn't help much because it's always set to a
random value, whether the user is logged in or not. As such,
the frontend needs a way of its own to store and read this
info.

How?
- Create a dedicated cookie ('fontLoggedIn') and set it to
  `true` when the user successfully logs in. Use the library
  'js-cookie' for that.
- Read this information in the Navbar, render differently
  according to this value.
- Because the app currently lacks a global state system, the
  page/app is simply reloaded to allow the Navbar to read
  again the cookie value and render properly. (Not super
  optimal for UX, but works.)